move model loading to utils
